# backend/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CompareRequest": {
        "properties": {
          "listingIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 4,
            "minItems": 1,
            "title": "Listingids",
            "type": "array"
          }
        },
        "required": [
          "listingIds"
        ],
        "title": "CompareRequest",
        "type": "object"
      },
      "CompareResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Listing"
            },
            "title": "Items",
            "type": "array"
          },
          "missingIds": {
            "items": {
              "type": "string"
            },
            "title": "Missingids",
            "type": "array"
          }
        },
        "required": [
          "items",
          "missingIds"
        ],
        "title": "CompareResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Listing": {
        "properties": {
          "bodyType": {
            "enum": [
              "SUV",
              "Sedan",
              "Coupe",
              "Hatchback",
              "Pickup",
              "Convertible"
            ],
            "title": "Bodytype",
            "type": "string"
          },
          "city": {
            "enum": [
              "Dubai",
              "Abu Dhabi",
              "Sharjah",
              "Ajman"
            ],
            "title": "City",
            "type": "string"
          },
          "conditionEvidence": {
            "items": {
              "$ref": "#/components/schemas/ConditionEvidence"
            },
            "title": "Conditionevidence",
            "type": "array"
          },
          "currency": {
            "const": "AED",
            "title": "Currency",
            "type": "string"
          },
          "dealPosition": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DealPosition"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "duplicateOffers": {
            "items": {
              "$ref": "#/components/schemas/DuplicateOffer"
            },
            "title": "Duplicateoffers",
            "type": "array"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "title": "Features",
            "type": "array"
          },
          "firstSeenAt": {
            "format": "date-time",
            "title": "Firstseenat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lastSeenAt": {
            "format": "date-time",
            "title": "Lastseenat",
            "type": "string"
          },
          "make": {
            "title": "Make",
            "type": "string"
          },
          "market": {
            "const": "ae",
            "title": "Market",
            "type": "string"
          },
          "mileageKm": {
            "minimum": 0.0,
            "title": "Mileagekm",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "photos": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Photos",
            "type": "array"
          },
          "price": {
            "minimum": 0.0,
            "title": "Price",
            "type": "integer"
          },
          "priceHistory": {
            "items": {
              "$ref": "#/components/schemas/PriceObservation"
            },
            "title": "Pricehistory",
            "type": "array"
          },
          "sellerType": {
            "enum": [
              "Dealer",
              "Private"
            ],
            "title": "Sellertype",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceAttribution"
          },
          "specifications": {
            "enum": [
              "GCC",
              "American",
              "European",
              "Japanese"
            ],
            "title": "Specifications",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trim": {
            "title": "Trim",
            "type": "string"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "market",
          "title",
          "make",
          "model",
          "trim",
          "year",
          "price",
          "currency",
          "mileageKm",
          "city",
          "bodyType",
          "specifications",
          "sellerType",
          "source",
          "description",
          "features",
          "photos",
          "firstSeenAt",
          "lastSeenAt",
          "conditionEvidence",
          "priceHistory",
          "duplicateOffers"
        ],
        "title": "Listing",
        "type": "object"
      },
      "ListingPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Listing"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageCount": {
            "title": "Pagecount",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "pageCount"
        ],
        "title": "ListingPage",
        "type": "object"
      },
...
      "ModelInsights": {
        "properties": {
          "listings": {
            "items": {
              "$ref": "#/components/schemas/Listing"
            },
            "title": "Listings",
            "type": "array"
          },
          "summary": {
            "$ref": "#/components/schemas/ModelMarketSummary"
          }
        },
        "required": [
          "summary",
          "listings"
        ],
        "title": "ModelInsights",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Carveo Catalogue API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/compare": {
      "post": {
        "operationId": "compare_api_v1_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompareResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/api/v1/listings": {
      "get": {
        "operationId": "listings_api_v1_listings_get",
        "parameters": [
          {
            "in": "query",
            "name": "market",
            "required": false,
            "schema": {
              "default": "ae",
              "title": "Market",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "make",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Make",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Model",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "bodyType",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Bodytype",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "yearMin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yearmin"
            }
          },
          {
            "in": "query",
            "name": "yearMax",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Yearmax"
            }
          },
          {
            "in": "query",
            "name": "priceMin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pricemin"
            }
          },
          {
            "in": "query",
            "name": "priceMax",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pricemax"
            }
          },
          {
            "in": "query",
            "name": "mileageMax",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mileagemax"
            }
          },
          {
            "in": "query",
            "name": "specifications",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Specifications",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "sellerType",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Sellertype",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "evidence",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Evidence",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "city",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "City",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "best-match",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 12,
              "title": "Pagesize",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listings",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/api/v1/listings/{listing_id}": {
      "get": {
        "operationId": "detail_api_v1_listings__listing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Listing"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detail",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/related": {
      "get": {
        "operationId": "related_api_v1_listings__listing_id__related_get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "title": "Listing Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 4,
              "maximum": 12,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Listing"
                  },
                  "title": "Response Related Api V1 Listings  Listing Id  Related Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Related",
        "tags": [
          "catalogue"
        ]
      }
    },
...
    "/api/v1/media/{photo_id}": {
      "get": {
        "operationId": "media_api_v1_media__photo_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Photo Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "image/gif": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/jpeg": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/png": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              },
              "image/webp": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Cached listing image"
          },
          "304": {
            "description": "Not modified"
          },
          "404": {
            "description": "Photo not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Media",
        "tags": [
          "catalogue"
        ]
      }
    },
    "/api/v1/models/{make}/{model}/insights": {
      "get": {
        "operationId": "insights_api_v1_models__make___model__insights_get",
        "parameters": [
          {
            "in": "path",
            "name": "make",
            "required": true,
            "schema": {
              "title": "Make",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "model",
            "required": true,
            "schema": {
              "title": "Model",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelInsights"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Insights",
        "tags": [
          "catalogue"
        ]
      }
    },
...
  }
}